global device list and stream
